Add a context manager for blocking signals
This makes the code cleaner

This should have been committed directly to master but since I am going to merge the i18n branch this evening, it does not really matters.

This context manager was badly needed for a plugins I am working on at work (a game editor plugin for our own game engine).