Optimize sql and add small sleeps between each connection to the db
This mitigate "OperationError: database locked" (the frontend is now able to read database while an indexing operation is running without facing the database locked issue) on slow systems (without ssd).

The main optimization was to move as much insert as possible in the same transation (and thus minimizing the number of transactions to perform).